requirements.txt
python-dotenv
firebase_admin
flask
requests
pandas